pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "fixture-fetcher"
version = "0.3.0"
description = "A Python CLI tool that fetches football fixture data and generates iCalendar files"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "rdawebb"}]
dependencies = [
    "beaupy>=3.11.0,<4.0.0",
    "icalendar>=6.3.2,<7.0.0",
    "pyyaml>=6.0.3,<7.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "requests>=2.32.5,<3.0.0",
    "rich>=14.2.0,<15.0.0",
]

[project.optional-dependencies]
dev = [
    "rust-just>=1.45,<2.0.0",
    "prek>=0.2.27,<1.0.0",
    "pytest>=9.0.2,<10.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-mock>=3.15.1,<4.0.0",
    "ruff>=0.14.11,<15.0.0",
    "ty>=0.0.11,<0.1.0"
]

[tool.ruff]
lint.ignore = [
    "B008",   # Do not perform function calls in argument defaults (common Typer pattern)
]
